requirements.txt
streamlit==1.31.1
google-cloud-bigquery==3.14.1
pandas==2.2.0
plotly==5.18.0
python-dotenv==1.0.1
db-dtypes==1.2.0 